+++ tori/centre.py
@@ -1,4 +1,5 @@
 from imagination.locator import Locator as ImaginationLocator
 
+core = None
 settings = {}
 services = ImaginationLocator()