+++ tori/controller.py
@@ -16,7 +16,7 @@ from re import match, sub
 from jinja2.exceptions import TemplateNotFound
 from tornado.web import HTTPError, RequestHandler
 
-from tori.centre             import services
+from tori.centre             import services, core
 from tori.common             import get_logger
 from tori.data.base          import ResourceEntity, resolve_file_path
 from tori.exception          import *
@@ -104,19 +104,10 @@ class Controller(RequestHandler):
         return 'cookie_secret' in self.settings and self.settings['cookie_secret']
 
     def render_template(self, template_name, **contexts):
+        """ Render the template with the given contexts.
+
+            See :meth:`tori.renderer.Renderer.render` for more information on the parameters.
         """
-        Render the template with the given contexts.
-
-        See :meth:`tori.renderer.Renderer.render` for more information.
-        """
-
-        # If the rendering source isn't set, break the code.
-        if not self._template_base_path:
-            raise RenderingSourceMissingError('The source of template is not identified. This method is disabled.')
-
-        # If the rendering engine is not specified, use the default one.
-        if not self._template_engine:
-            self._template_engine = DefaultRenderer
 
         contexts['app'] = {
             'request': self.request,
@@ -131,10 +122,9 @@ class Controller(RequestHandler):
         return output
 
     def render(self, template_name, **contexts):
-        """
-        Render the template with the given contexts and push the output buffer.
+        """ Render the template with the given contexts and push the output buffer.
 
-        See :meth:`tori.renderer.Renderer.render` for more information.
+            See :meth:`tori.renderer.Renderer.render` for more information on the parameters.
         """
         self.write(self.render_template(template_name, **contexts))
 
@@ -142,8 +132,25 @@ class Controller(RequestHandler):
     def template_engine(self):
         """ Template Engine
 
-        :rtype: tori.template.renderer.Renderer
+            :rtype: tori.template.renderer.Renderer
+            :raise RenderingSourceMissingError: if the template base path and the reverse core reference are not defined.
+
+            .. versionchanged:: 3.0
+
+                The exception will not be raised if the reverse core reference is defined.
+
+            .. note::
+
+                The reverse core reference is the first instance of
+                :class:`tori.application.Application` in the process.
         """
+
+        # If the rendering source isn't set, break the code.
+        if not centre.core and not self._template_base_path:
+            raise RenderingSourceMissingError('The source of template is not identified. This method is disabled.')
+
+        if not self._template_base_path:
+            self._template_base_path = p.join(centre.core.base_path, 'templates')
 
         # If the rendering engine is not specified, use the default one.
         if not self._template_engine: