+++ docs/source/changes.rst
@@ -17,7 +17,7 @@ Version 3.1
 Version 3.0
 ===========
 
-:Release Date: 2014
+:Release Date: 2014.08.16
 
 - **ORM/tori.db**: Allow cross-collection (or cross-repository) queries within the same type of backend datastores.
 - **ORM/tori.db**: (**BCB-2.1**) Removed the silly preconditions of the setup of ORM.
@@ -29,4 +29,7 @@ Version 3.0
   as **Criteria**. This change is to address the semantic / readability issue. (Hence, all references to Criteria objects
   are now referred to Query objects.)
 - **ORM/tori.db**: Removed unused / tedious code from the ORM.
+- **Web Framework**: the first instance of :class:`tori.application.Application` is now self-referenced as ``tori.centre.core``.
+- **Web Framework**: Without specifying the rendering path for each controller, the controller will be looking for
+  templates from :file:`<app_base_path>/templates`.
 - **Tests**: Reorganized the tests and refactored the ORM tests.